Move projectdirectory to zip to messages
